ethereumetl/cli package
(1) original scripts in root directory are moved to `ethereumetl/cli`
(2) `main()` is renamed to corresponding script names
(3) root scripts invoke functions in `ethereumetl/cli`
