+++ ethereumetl/cli/get_keccak_hash.py
@@ -21,6 +21,20 @@
 # SOFTWARE.
 
 
-from ethereumetl.cli.extract_token_transfers import extract_token_transfers
+import click
 
-extract_token_transfers()
+from eth_utils import keccak
+
+from ethereumetl.file_utils import smart_open
+from ethereumetl.logging_utils import logging_basic_config
+
+@click.command(context_settings=dict(help_option_names=['-h', '--help']))
+@click.option('-i', '--input-string', default='Transfer(address,address,uint256)', type=str, help='String to hash, e.g. Transfer(address,address,uint256)')
+@click.option('-o', '--output', default='-', type=str, help='The output file. If not specified stdout is used.')
+
+def get_keccak_hash(input_string, output):
+    """Outputs the 32-byte keccak hash of the given string."""
+    hash = keccak(text=input_string)
+
+    with smart_open(output, 'w') as output_file:
+        output_file.write('0x{}\n'.format(hash.hex()))
