+++ export_contracts.py
@@ -21,37 +21,6 @@
 # SOFTWARE.
 
 
-import click
+from ethereumetl.cli.export_contracts import export_contracts
 
-from ethereumetl.file_utils import smart_open
-from ethereumetl.jobs.export_contracts_job import ExportContractsJob
-from ethereumetl.jobs.exporters.contracts_item_exporter import contracts_item_exporter
-from ethereumetl.logging_utils import logging_basic_config
-from ethereumetl.thread_local_proxy import ThreadLocalProxy
-from ethereumetl.providers.auto import get_provider_from_uri
-
-logging_basic_config()
-
-@click.command(context_settings=dict(help_option_names=['-h', '--help']))
-@click.option('-b', '--batch-size', default=100, type=int, help='The number of blocks to filter at a time.')
-@click.option('-c', '--contract-addresses', type=str, help='The file containing contract addresses, one per line.')
-@click.option('-o', '--output', default='-', type=str, help='The output file. If not specified stdout is used.')
-@click.option('-w', '--max-workers', default=5, type=int, help='The maximum number of workers.')
-@click.option('-p', '--provider-uri', default='https://mainnet.infura.io', type=str, help='The URI of the web3 provider e.g. file://$HOME/Library/Ethereum/geth.ipc or https://mainnet.infura.io')
-
-def main(batch_size, contract_addresses, output, max_workers, provider_uri):
-    """Exports contracts bytecode using eth_getCode JSON RPC APIs."""
-    with smart_open(contract_addresses, 'r') as contract_addresses_file:
-        contract_addresses = (contract_address.strip() for contract_address in contract_addresses_file
-                              if contract_address.strip())
-        job = ExportContractsJob(
-            contract_addresses_iterable=contract_addresses,
-            batch_size=batch_size,
-            batch_web3_provider=ThreadLocalProxy(lambda: get_provider_from_uri(provider_uri, batch=True)),
-            item_exporter=contracts_item_exporter(output),
-            max_workers=max_workers)
-
-        job.run()
-
-if __name__ == '__main__':
-    main()
+export_contracts()
