+++ export_receipts_and_logs.py
@@ -21,38 +21,6 @@
 # SOFTWARE.
 
 
-import click
+from ethereumetl.cli.export_receipts_and_logs import export_receipts_and_logs
 
-from ethereumetl.file_utils import smart_open
-from ethereumetl.jobs.export_receipts_job import ExportReceiptsJob
-from ethereumetl.jobs.exporters.receipts_and_logs_item_exporter import receipts_and_logs_item_exporter
-from ethereumetl.logging_utils import logging_basic_config
-from ethereumetl.thread_local_proxy import ThreadLocalProxy
-from ethereumetl.providers.auto import get_provider_from_uri
-
-logging_basic_config()
-
-@click.command(context_settings=dict(help_option_names=['-h', '--help']))
-@click.option('-b', '--batch-size', default=100, type=int, help='The number of receipts to export at a time.')
-@click.option('-t', '--transaction-hashes', type=str, help='The file containing transaction hashes, one per line.')
-@click.option('-p', '--provider-uri', default='https://mainnet.infura.io', type=str, help='The URI of the web3 provider e.g. file://$HOME/Library/Ethereum/geth.ipc or https://mainnet.infura.io')
-@click.option('-w', '--max-workers', default=5, type=int, help='The maximum number of workers.')
-@click.option('--receipts-output', default=None, type=str, help='The output file for receipts. If not provided receipts will not be exported. Use "-" for stdout')
-@click.option('--logs-output', default=None, type=str, help='The output file for receipt logs. If not provided receipt logs will not be exported. Use "-" for stdout')
-
-def main(batch_size, transaction_hashes, provider_uri, max_workers, receipts_output, logs_output):
-    """Export receipts and logs."""
-    with smart_open(transaction_hashes, 'r') as transaction_hashes_file:
-        job = ExportReceiptsJob(
-            transaction_hashes_iterable=(transaction_hash.strip() for transaction_hash in transaction_hashes_file),
-            batch_size=batch_size,
-            batch_web3_provider=ThreadLocalProxy(lambda: get_provider_from_uri(provider_uri, batch=True)),
-            max_workers=max_workers,
-            item_exporter=receipts_and_logs_item_exporter(receipts_output, logs_output),
-            export_receipts=receipts_output is not None,
-            export_logs=logs_output is not None)
-
-        job.run()
-
-if __name__ == '__main__':
-    main()
+export_receipts_and_logs()
