+++ export_token_transfers.py
@@ -21,38 +21,6 @@
 # SOFTWARE.
 
 
-import click
+from ethereumetl.cli.export_token_transfers import export_token_transfers
 
-from web3 import Web3
-
-from ethereumetl.jobs.export_token_transfers_job import ExportTokenTransfersJob
-from ethereumetl.jobs.exporters.token_transfers_item_exporter import token_transfers_item_exporter
-from ethereumetl.logging_utils import logging_basic_config
-from ethereumetl.providers.auto import get_provider_from_uri
-from ethereumetl.thread_local_proxy import ThreadLocalProxy
-
-logging_basic_config()
-
-@click.command(context_settings=dict(help_option_names=['-h', '--help']))
-@click.option('-s', '--start-block', default=0, type=int, help='Start block')
-@click.option('-e', '--end-block', required=True, type=int, help='End block')
-@click.option('-b', '--batch-size', default=100, type=int, help='The number of blocks to filter at a time.')
-@click.option('-o', '--output', default='-', type=str, help='The output file. If not specified stdout is used.')
-@click.option('-w', '--max-workers', default=5, type=int, help='The maximum number of workers.')
-@click.option('-p', '--provider-uri', required=True, type=str, help='The URI of the web3 provider e.g. file://$HOME/Library/Ethereum/geth.ipc or http://localhost:8545/')
-@click.option('-t', '--tokens', default=None, type=str, nargs=1, help='The list of token addresses to filter by.')
-
-def main(start_block, end_block, batch_size, output, max_workers, provider_uri, tokens):
-    """Exports ERC20 transfers using eth_newFilter and eth_getFilterLogs JSON RPC APIs."""
-    job = ExportTokenTransfersJob(
-        start_block=start_block,
-        end_block=end_block,
-        batch_size=batch_size,
-        web3=ThreadLocalProxy(lambda: Web3(get_provider_from_uri(provider_uri))),
-        item_exporter=token_transfers_item_exporter(output),
-        max_workers=max_workers,
-        tokens=tokens)
-    job.run()
-
-if __name__ == '__main__':
-    main()
+export_token_transfers()
